requirements.txt
faiss-cpu
google-ai-generativelanguage
google-generativeai
langchain
langchain-community
PyPDF2
scikit-learn
sentence-transformers
spacy
streamlit